Use this one to get working expert demos
prior to 12/5/2025 5:20pm all the tests using demo_data folder trajectories based on my pytorch ones were using trajectories made by random actions because the generate_demos.py code was using its random action fallback because of how the code was made, it had no non fall back, so I made generate_demos_new.py it can load and rebuild the demos, (actually it might even be possible to generate one on your machine and to render it while it goes). This is the file that uses the ones made by this code, since this version expects to be given arguments and to only do one at a time so this has a for loop in it that enables it to do all of them, the 2nd commented out version is a simple one that works.