Working expert demonstration trajectories
Each at 10 episodes with 1000 steps these are the only ones necessary for our purposes, the "expert" ones become the new top and the medium are still the mid, this is that change I refered to being done at 12/5/2025 5:20 pm but it took me a while to get to it but now I have, so with the generate_demos_new.py that one created these and was actually able to run the expert demo files (instead of just using random actions, also it might even be possible to render it while it does this but I have not tried that yet). This gets ran using the run_experts_new.sh file that loops its calls of generate_demos_new.py to get the necessary trajectories. And these have been tested in terms of their format and are 100% compatible with the original pytorch version code.